+++ src/components/ProfilePage/ProfilePage.js
@@ -9,11 +9,6 @@ import {
   Menu,
   Select,
   MenuItem,
-  TextField,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogContentText,
   CircularProgress
 } from '@material-ui/core';
 import styled from 'styled-components';
@@ -31,19 +26,26 @@ import {
   typeList
 } from '../../DataExport';
 
+const MainWrapper = styled.div`
+  margin: auto;
+  margin-top: 3rem;
+  width: 70vw;
+  padding: 1.5rem;
+  display: display;
+  align-items: center;
+  background-color: #ffffff;
+  box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
+  @media ${props => props.theme.mobile} {
+    width: calc(100vw - 1rem);
+    margin-top: 1rem;
+    padding: 0.5rem;
+  }
+`;
+
 const Wrapper = styled.div`
   width: 100%;
   display: flex;
   justify-content: center;
-`;
-
-const InfoTableWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  max-width: 396px;
 `;
 
 const InfoTable = styled(TableContainer)`
@@ -91,19 +93,6 @@ const FormInput = styled.input`
   }
 `;
 
-const FormField = styled.form`
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  width: 100%;
-  align-items: center;
-`;
-
-const FormTextField = styled(TextField)`
-  margin: 12px 32px 12px 32px;
-  min-width: 256px;
-`;
-
 const ErrorText = styled.div`
   margin: 6px;
   font-size: 18px;
@@ -113,18 +102,6 @@ const ErrorText = styled.div`
 const FormSelectField = styled(Select)`
   margin-right: 2px;
   width: 128px;
-`;
-
-const ButtonWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-`;
-
-const HalfButton = styled(Button)`
-  width: 50%;
-  padding: 8px;
-  margin: 8px;
 `;
 
 const RefreshIcon = styled(FontAwesomeIcon)`
@@ -142,126 +119,14 @@ const LoadingCircle = styled(CircularProgress)`
   bottom: 50vh;
 `;
 
-const PwChangeDialog = props => {
-  const { mode, error, newPwForm, modeChange, pwChangeSubmit, formChange } =
-    props;
-  return (
-    <>
-      <Dialog
-        open={mode === 'pwChangeMode'}
-        onClose={modeChange}
-        aria-labelledby="form-dialog-title"
-      >
-        <DialogTitle id="form-dialog-title">비밀번호 변경</DialogTitle>
-        <FormField onSubmit={pwChangeSubmit}>
-          <FormTextField
-            autoFocus
-            id="nowPassword"
-            label="현재 비밀번호를 입력하세요"
-            type="password"
-            onChange={formChange}
-            value={newPwForm.nowPassword}
-          />
-          <FormTextField
-            id="newPassword"
-            label="새로운 비밀번호"
-            type="password"
-            onChange={formChange}
-            value={newPwForm.newPassword}
-          />
-          <FormTextField
-            id="newPasswordConfirm"
-            label="새로운 비밀번호 확인"
-            type="password"
-            onChange={formChange}
-            value={newPwForm.newPasswordConfirm}
-          />
-          <ErrorText>{error}</ErrorText>
-          <ButtonWrapper>
-            <HalfButton onClick={modeChange} color="primary">
-              취소
-            </HalfButton>
-            <HalfButton type="submit" onClick={pwChangeSubmit} color="primary">
-              변경하기
-            </HalfButton>
-          </ButtonWrapper>
-        </FormField>
-      </Dialog>
-    </>
-  );
-};
-
-const WithdrawalDialog = props => {
-  const {
-    mode,
-    error,
-    withDrawalForm,
-    modeChange,
-    formChange,
-    withdrawalSubmit
-  } = props;
-  return (
-    <>
-      <Dialog
-        open={mode === 'withdrawalMode'}
-        onClose={modeChange}
-        aria-labelledby="form-dialog-title"
-      >
-        <DialogTitle id="form-dialog-title">회원 탈퇴</DialogTitle>
-        <DialogContent>
-          <DialogContentText> 나 탈퇴한다~~~~~~~~~</DialogContentText>
-        </DialogContent>
-        <FormField submit={withdrawalSubmit}>
-          <FormTextField
-            autoFocus
-            id="password"
-            label="비밀번호를 입력하세요"
-            type="password"
-            onChange={formChange}
-            value={withDrawalForm.password}
-          />
-          <FormTextField
-            id="text"
-            label="'탈퇴'를 입력하세요"
-            type="text"
-            onChange={formChange}
-            value={withDrawalForm.text}
-          />
-          <Wrapper>
-            <ErrorText style={{ fontSize: '16px' }}>{error}</ErrorText>
-          </Wrapper>
-          <ButtonWrapper>
-            <HalfButton onClick={modeChange} color="primary">
-              취소
-            </HalfButton>
-            <HalfButton
-              type="submit"
-              onClick={withdrawalSubmit}
-              color="primary"
-            >
-              탈퇴
-            </HalfButton>
-          </ButtonWrapper>
-        </FormField>
-      </Dialog>
-    </>
-  );
-};
-
 const ProfileHeader = props => {
   const {
     mode,
     anchorEl,
-    error,
-    newPwForm,
-    withDrawalForm,
     profileUserId,
     userId,
     modeChange,
     menuClick,
-    formChange,
-    pwChangeSubmit,
-    withdrawalSubmit,
     editFormRefresh
   } = props;
 
@@ -319,22 +184,6 @@ const ProfileHeader = props => {
           </Menu>
         </>
       )}
-      <PwChangeDialog
-        mode={mode}
-        error={error}
-        newPwForm={newPwForm}
-        modeChange={modeChange}
-        formChange={formChange}
-        pwChangeSubmit={pwChangeSubmit}
-      />
-      <WithdrawalDialog
-        mode={mode}
-        withDrawalForm={withDrawalForm}
-        error={error}
-        modeChange={modeChange}
-        formChange={formChange}
-        withdrawalSubmit={withdrawalSubmit}
-      />
     </MyinfoHeader>
   );
 };
@@ -497,16 +346,8 @@ const ProfileBody = props => {
 };
 
 const PropfilePage = props => {
-  const {
-    infoObj,
-    infoEditObj,
-    anchorEl,
-    newPwForm,
-    error,
-    withDrawalForm,
-    mode,
-    myinformationLoading
-  } = props;
+  const { infoObj, infoEditObj, anchorEl, error, mode, myinformationLoading } =
+    props;
 
   const {
     handleChange,
@@ -515,9 +356,7 @@ const PropfilePage = props => {
     myinfoEditSubmit,
     menuClick,
     formChange,
-    pwChangeSubmit,
     typeChange,
-    withdrawalSubmit,
     editFormRefresh
   } = props;
 
@@ -532,35 +371,28 @@ const PropfilePage = props => {
   const profileUserId = infoObj.idString;
   const userId = localStorage.getItem('userId');
   return (
-    <Wrapper>
-      <InfoTableWrapper>
-        <ProfileHeader
-          mode={mode}
-          error={error}
-          anchorEl={anchorEl}
-          newPwForm={newPwForm}
-          withDrawalForm={withDrawalForm}
-          profileUserId={profileUserId}
-          userId={userId}
-          formChange={formChange}
-          modeChange={modeChange}
-          menuClick={menuClick}
-          pwChangeSubmit={pwChangeSubmit}
-          withdrawalSubmit={withdrawalSubmit}
-          editFormRefresh={editFormRefresh}
-        />
-        <ProfileBody
-          rows={rows}
-          infoEditObj={infoEditObj}
-          informationOpenAgreeChange={informationOpenAgreeChange}
-          handleChange={handleChange}
-          typeChange={typeChange}
-          mode={mode}
-          myinfoEditSubmit={myinfoEditSubmit}
-        />
-        <ErrorText>{error}</ErrorText>
-      </InfoTableWrapper>
-    </Wrapper>
+    <MainWrapper>
+      <ProfileHeader
+        mode={mode}
+        anchorEl={anchorEl}
+        profileUserId={profileUserId}
+        userId={userId}
+        formChange={formChange}
+        modeChange={modeChange}
+        menuClick={menuClick}
+        editFormRefresh={editFormRefresh}
+      />
+      <ProfileBody
+        rows={rows}
+        infoEditObj={infoEditObj}
+        informationOpenAgreeChange={informationOpenAgreeChange}
+        handleChange={handleChange}
+        typeChange={typeChange}
+        mode={mode}
+        myinfoEditSubmit={myinfoEditSubmit}
+      />
+      <ErrorText>{error}</ErrorText>
+    </MainWrapper>
   );
 };
